requirements.txt
flask
aiohttp
botbuilder-core
botbuilder-schema
python-dotenv
pyodbc
